Fix tests and added :region
