+++ endpoints/road/tests/integration_test.js
@@ -1,16 +1,7 @@
 const request = require('request')
 const helpers = require('../../../lib/test_helpers')
 
-describe('road root', () => {
-  it('should return info', (done) => {
-    const fieldsToCheckFor = ['info']
-    const params = helpers.testRequestParams('/road/')
-    const resultHandler = helpers.testRequestHandlerForFields(done, fieldsToCheckFor)
-    request.get(params, resultHandler)
-  })
-})
-
-describe('road - all', () => {
+describe('road', () => {
   it('should return an array of objects containing correct fields', (done) => {
     const fieldsToCheckFor = [
       'routeId',
@@ -31,7 +22,7 @@ describe('road - all', () => {
       'conditionUpdated',
       'surfaceCondition',
     ]
-    const params = helpers.testRequestParams('/road/all')
+    const params = helpers.testRequestParams('/road')
     const resultHandler = helpers.testRequestHandlerForFields(done, fieldsToCheckFor)
     request.get(params, resultHandler)
   })
